Re-enable job queries and update design of worker detail
