+++ crates/hyperqueue/src/dashboard/ui/screens/cluster_overview/worker/cpu_util_table.rs
@@ -21,9 +21,13 @@ pub fn render_cpu_util_table(
     mem_util: &MemoryStats,
     rect: Rect,
     frame: &mut DashboardFrame,
-    constraints: &[Constraint],
     table_style: Style,
 ) {
+    if cpu_util_list.is_empty() {
+        return;
+    }
+    let constraints = get_column_constraints(rect, cpu_util_list.len());
+
     let width = constraints.len();
     let height = (cpu_util_list.len() as f64 / width as f64).ceil() as usize;
 
@@ -81,7 +85,7 @@ pub fn render_cpu_util_table(
 }
 
 /// Creates the column sizes for the cpu_util_table, each column divides the row equally.
-pub fn get_column_constraints(rect: Rect, num_cpus: usize) -> Vec<Constraint> {
+fn get_column_constraints(rect: Rect, num_cpus: usize) -> Vec<Constraint> {
     let max_columns = (rect.width / CPU_METER_WIDTH as u16) as usize;
     let num_columns = cmp::min(max_columns, num_cpus);
